requirements.txt
boto3~=1.26.104
pandas~=1.5.3
plotly~=5.14.1
numpy~=1.24.2
constructs~=10.1.299
aws-cdk-lib~=2.72.1
cdk-nag~=2.25.2
dynamo-pandas~=1.3.0
tables~=3.8.0
kaleido==0.2.1; sys_platform == 'linux'
kaleido==0.1.0post1; sys_platform == 'win32'
XlsxWriter~=3.1.2
